+++ lib/paginate_firestore.dart
@@ -24,7 +24,7 @@ class PaginateFirestore extends StatefulWidget {
     required this.query,
     required this.itemBuilderType,
     this.gridDelegate =
-        const SliverGridDelegateWithFixedCrossAxisCount(crossAxisCount: 2),
+    const SliverGridDelegateWithFixedCrossAxisCount(crossAxisCount: 2),
     this.startAfterDocument,
     this.itemsPerPage = 15,
     this.onError,
@@ -138,8 +138,8 @@ class _PaginateFirestoreState extends State<PaginateFirestore> {
           return widget.itemBuilderType == PaginateBuilderType.listView
               ? _buildListView(loadedState)
               : widget.itemBuilderType == PaginateBuilderType.gridView
-                  ? _buildGridView(loadedState)
-                  : _buildPageView(loadedState);
+              ? _buildGridView(loadedState)
+              : _buildPageView(loadedState);
         }
       },
     );
@@ -165,13 +165,14 @@ class _PaginateFirestoreState extends State<PaginateFirestore> {
   @override
   void initState() {
     customRefresher = CustomRefresher(
-      enablePullDown: widget.customRefresher?.enablePullDown,
-      enablePullUp: widget.customRefresher?.enablePullUp,
-      enableTwoLevel: widget.customRefresher?.enableTwoLevel,
-      footer: widget.customRefresher?.footer,
-      header: widget.customRefresher?.header,
-      onRefresh: widget.customRefresher!.onRefresh,
-      physics: widget.physics
+        enablePullDown: widget.customRefresher?.enablePullDown,
+        enablePullUp: widget.customRefresher?.enablePullUp,
+        enableTwoLevel: widget.customRefresher?.enableTwoLevel,
+        footer: widget.customRefresher?.footer,
+        header: widget.customRefresher?.header,
+        onRefresh: widget.customRefresher?.onRefresh,
+        physics: widget.physics,
+        refreshController: widget.customRefresher?.refreshController ?? refreshController
     );
 
     if (widget.listeners != null) {
@@ -217,7 +218,7 @@ class _PaginateFirestoreState extends State<PaginateFirestore> {
           sliver: SliverGrid(
             gridDelegate: widget.gridDelegate,
             delegate: SliverChildBuilderDelegate(
-              (context, index) {
+                  (context, index) {
                 if (index >= loadedState.documentSnapshots.length) {
                   _cubit!.fetchPaginatedList();
                   return widget.bottomLoader;
@@ -242,8 +243,8 @@ class _PaginateFirestoreState extends State<PaginateFirestore> {
       return MultiProvider(
         providers: widget.listeners!
             .map((listener) => ChangeNotifierProvider(
-                  create: (context) => listener,
-                ))
+          create: (context) => listener,
+        ))
             .toList(),
         child: gridView,
       );
@@ -252,7 +253,6 @@ class _PaginateFirestoreState extends State<PaginateFirestore> {
     if(widget.hasRefreshIndicator == true) {
       return RefreshIndicatorWidget(
         scrollController: widget.scrollController!,
-        refreshController: refreshController,
         customRefresher: customRefresher!,
         child: gridView,
       );
@@ -275,7 +275,7 @@ class _PaginateFirestoreState extends State<PaginateFirestore> {
           padding: widget.padding,
           sliver: SliverList(
             delegate: SliverChildBuilderDelegate(
-              (context, index) {
+                  (context, index) {
                 final itemIndex = index ~/ 2;
                 if (index.isEven) {
                   if (itemIndex >= loadedState.documentSnapshots.length) {
@@ -300,9 +300,9 @@ class _PaginateFirestoreState extends State<PaginateFirestore> {
               childCount: max(
                   0,
                   (loadedState.hasReachedEnd
-                              ? loadedState.documentSnapshots.length
-                              : loadedState.documentSnapshots.length + 1) *
-                          2 -
+                      ? loadedState.documentSnapshots.length
+                      : loadedState.documentSnapshots.length + 1) *
+                      2 -
                       1),
             ),
           ),
@@ -315,8 +315,8 @@ class _PaginateFirestoreState extends State<PaginateFirestore> {
       return MultiProvider(
         providers: widget.listeners!
             .map((listener) => ChangeNotifierProvider(
-                  create: (context) => listener,
-                ))
+          create: (context) => listener,
+        ))
             .toList(),
         child: listView,
       );
@@ -325,7 +325,6 @@ class _PaginateFirestoreState extends State<PaginateFirestore> {
     if(widget.hasRefreshIndicator == true) {
       return RefreshIndicatorWidget(
         scrollController: widget.scrollController!,
-        refreshController: refreshController,
         customRefresher: customRefresher!,
         child: listView,
       );
@@ -345,7 +344,7 @@ class _PaginateFirestoreState extends State<PaginateFirestore> {
         physics: widget.physics,
         onPageChanged: widget.onPageChanged,
         childrenDelegate: SliverChildBuilderDelegate(
-          (context, index) {
+              (context, index) {
             if (index >= loadedState.documentSnapshots.length) {
               _cubit!.fetchPaginatedList();
               return widget.bottomLoader;
@@ -367,8 +366,8 @@ class _PaginateFirestoreState extends State<PaginateFirestore> {
       return MultiProvider(
         providers: widget.listeners!
             .map((listener) => ChangeNotifierProvider(
-                  create: (context) => listener,
-                ))
+          create: (context) => listener,
+        ))
             .toList(),
         child: pageView,
       );
@@ -377,7 +376,6 @@ class _PaginateFirestoreState extends State<PaginateFirestore> {
     if(widget.hasRefreshIndicator == true) {
       return RefreshIndicatorWidget(
         scrollController: widget.scrollController!,
-        refreshController: refreshController,
         customRefresher: customRefresher!,
         child: pageView,
       );